refactor: move db di module directory
